Added baseline version of MATfluids
This is the first upload of the MATfluids package (Part 1). There is much development (and standardizing) to be done and a manual to be written, but the library is functional as it is and contains examples within the code itself for the moment. Part 2 will arrive in one day. And so the development begins.